chore: Fix doctests compilation
This problem was introduced when `Button` and `Label` were becoming a
`<State>`ful type, and in the examples, the compiler could not infer the
generic type.

Usually, these types are constructed in a way, where the compiler knows
what `<State>` is, because the object is used in `make_main_window`.